Quick Rotate Buttons no longer need to be held and 45s are a bit easier
After simulation testing AbsoluteDriveAdv it need a bit of tuning as the robot turned the wrong direction. Also the quick rotate buttons would have to be held down previously as the joystick adjustment would overwrite them as it would always be changing with the gyro. I also changed the default constant as it was unbareably slow.