Heading correction logic now reads in current heading
The purpose of this commit is address the issue where heading correction is broken as indicated in the following link:
https://github.com/BroncBotz3481/YAGSL/issues/44

The cause of this issue seems to be from the heading controller using the lastHeadingRadians as its measurement and the current heading as its target, causing the output to be inverted. This commit fixes this by swapping these values.